split filter_function
